feat(api): add infrastructure specific config
Infrastructure scpecific config can be defined under the "infr" property
of the json.
This first itération of this feature currently uses a Map with enabled
additionalProperties. Later I would like to add a more strict version of
this feature where an Infrastructure would define a POJO containing
its config so that Jackson could automatically parse it.