perf(generic): save MessageHandlers in a Map
This commit adds the MessageDispatcher class and uses it in NodeProcess.
This way the NodeProcess is processed only once and the result is saved
in a Map so it can be quickly accessed on each call to processMessage.
Also it adds a check to make sure that there is only one handler for
each class of Message.