docs(white-paper): add white paper manually
I didn't use git to merge the rapport branch as it was too messy (300+ useless files)
so instead I decided to add the files manually. I also cleaned and renamed some files
as well as adding a Makefile to render the document and clean the workspace.

I also removed the docs section from the top level README and put it in a new README
in docs/white-paper-2020-06.